// src/app/models/model/nebim/cdCreditCardTypeDesc.ts
export class cdCreditCardTypeDesc {
  creditCardTypeCode: string;
  isCreditCardNumRequired: boolean;
  isCreditCardValidCheck: boolean;
  maxInstallmentCount: number;
  availableInstallments: string;
  likeCashInstallments: string;
  bankCode: string;
  bankCardTypeCode: number;
  isContractedBankCard: boolean;
  isDebitCard: boolean;
  isOverseasBankCard: boolean;
  isCompanyCreditCardType: boolean;
  editMaskCode: number;
  isBlocked: boolean;
  langCode: string;
  creditCardTypeDescription: string;
}